Address review comments: Clean up package.json, rename and remove some stuff
