+++ lib/apollo/initApollo.js
@@ -1,7 +1,7 @@
 import { ApolloClient, createNetworkInterface } from 'react-apollo'
 import { SubscriptionClient, addGraphQLSubscriptions } from 'subscriptions-transport-ws'
 import fetch from 'isomorphic-fetch'
-import {API_BASE_URL, API_WS_BASE_URL} from '../settings'
+import {API_BASE_URL, API_WS_BASE_URL} from '../constants'
 
 let apolloClient = null
