+++ lib/utils/name.js
@@ -1,3 +1,4 @@
+// TODO: Move to backend.
 export const getName = user => (
     (user.name && user.name.trim()) ||
     cleanName(user.email)
