+++ package.json
@@ -22,7 +22,7 @@
     "url": "git+https://github.com/orbiting/republik-frontend.git"
   },
   "author": "",
-  "license": "BSD-3",
+  "license": "BSD-3-Clause",
   "bugs": {
     "url": "https://github.com/orbiting/republik-frontend/issues"
   },
@@ -47,52 +47,31 @@
     ]
   },
   "devDependencies": {
-    "babel-plugin-transform-react-jsx": "^6.24.1",
-    "babel-preset-env": "^1.6.0",
     "babel-tape-runner": "^2.0.1",
-    "enzyme": "^2.9.1",
     "eslint-plugin-import": "^2.7.0",
     "faucet": "0.0.1",
     "gsheets": "^2.0.0",
-    "husky": "^0.14.3",
     "lint-staged": "^4.0.4",
     "nodemon": "^1.11.0",
-    "react-test-renderer": "^15.6.1",
-    "spy": "^1.0.0",
     "standard": "^10.0.3",
     "tape": "^4.8.0"
   },
   "dependencies": {
     "@project-r/styleguide": "^1.1.0",
-    "d3-array": "^1.2.0",
-    "d3-color": "^1.0.3",
     "d3-format": "^1.2.0",
-    "d3-scale": "^1.0.6",
     "d3-time-format": "^2.0.5",
     "dotenv": "^4.0.0",
     "email-validator": "^1.1.1",
     "express": "^4.15.4",
     "express-basic-auth": "^1.1.1",
     "glamor": "^2.20.39",
-    "immutable": "^3.8.1",
-    "is-url": "^1.2.2",
-    "js-yaml": "^3.9.1",
     "next": "^4.0.0-beta.2",
     "next-routes": "^1.0.40",
     "prop-types": "^15.5.10",
-    "ramda": "^0.24.1",
     "react": "^16.0.0",
     "react-apollo": "^1.4.15",
     "react-dom": "^16.0.0",
-    "react-icons": "^2.2.5",
-    "react-maskedinput": "^4.0.0",
-    "react-portal": "^3.1.0",
-    "react-textarea-autosize": "^5.1.0",
     "redux": "^3.7.2",
-    "remark-frontmatter": "^1.1.0",
-    "remark-parse": "^4.0.0",
-    "remark-stringify": "^4.0.0",
-    "subscriptions-transport-ws": "^0.8.2",
-    "unified": "^6.1.5"
+    "subscriptions-transport-ws": "^0.8.2"
   }
 }
